+++ packages/http/src/index.ts
@@ -59,5 +59,5 @@ export { timing } from './utils/timing';
 export { once } from './utils/promise';
 export { invariant } from './utils/invariant';
 export { TokenBucket, type TokenBucketOptions } from './utils/token-bucket';
-
+export { fetch, type RetryOptions } from './utils/fetch';
 export { ISO_3601_1, type ISO3166CountryCode } from './iso/iso_3601_1';