+++ packages/utils/eslint.config.js
@@ -1,0 +1,4 @@
+import { packageConfig } from '@repo/eslint-config/package';
+
+/** @type {import("eslint").Linter.Config} */
+export default packageConfig;